refactor: modify overview package structure
Move the overview packages out into the top level
