+++ app/src/main/java/com/chesire/malime/flow/LaunchActivity.kt
@@ -3,7 +3,6 @@ package com.chesire.malime.flow
 import android.content.Intent
 import android.os.Bundle
 import com.chesire.malime.flow.login.LoginActivity
-import com.chesire.malime.flow.overview.OverviewActivity
 import com.chesire.malime.kitsu.AuthProvider
 import dagger.android.DaggerActivity
 import javax.inject.Inject
