+++ app/src/main/java/com/chesire/malime/flow/activity/ActivityFragment.kt
@@ -1,27 +1,27 @@
-package com.chesire.malime.flow.overview.manga
+package com.chesire.malime.flow.activity
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.chesire.malime.databinding.FragmentMangaBinding
+import com.chesire.malime.databinding.FragmentActivityBinding
 import dagger.android.support.DaggerFragment
 
-class MangaFragment : DaggerFragment() {
+class ActivityFragment : DaggerFragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = FragmentMangaBinding.inflate(inflater, container, false).root
+    ) = FragmentActivityBinding.inflate(inflater, container, false).root
 
     companion object {
-        const val TAG = "MangaFragment"
+        const val TAG = "ActivityFragment"
 
         /**
          * Use this factory method to create a new instance of
          * this fragment using the provided parameters.
          *
-         * @return A new instance of fragment MangaFragment.
+         * @return A new instance of fragment ActivityFragment.
          */
-        fun newInstance() = MangaFragment()
+        fun newInstance() = ActivityFragment()
     }
 }
