+++ app/src/main/java/com/chesire/malime/flow/login/LoginActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.transaction
 import com.chesire.malime.R
-import com.chesire.malime.flow.overview.OverviewActivity
+import com.chesire.malime.flow.OverviewActivity
 
 class LoginActivity : AppCompatActivity(), LoginFragment.LoginListener {
     override fun onCreate(savedInstanceState: Bundle?) {
