+++ app/src/main/java/com/chesire/malime/flow/profile/ProfileFragment.kt
@@ -1,27 +1,27 @@
-package com.chesire.malime.flow.overview.activity
+package com.chesire.malime.flow.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.chesire.malime.databinding.FragmentActivityBinding
+import com.chesire.malime.databinding.FragmentProfileBinding
 import dagger.android.support.DaggerFragment
 
-class ActivityFragment : DaggerFragment() {
+class ProfileFragment : DaggerFragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = FragmentActivityBinding.inflate(inflater, container, false).root
+    ) = FragmentProfileBinding.inflate(inflater, container, false).root
 
     companion object {
-        const val TAG = "ActivityFragment"
+        const val TAG = "ProfileFragment"
 
         /**
          * Use this factory method to create a new instance of
          * this fragment using the provided parameters.
          *
-         * @return A new instance of fragment ActivityFragment.
+         * @return A new instance of fragment ProfileFragment.
          */
-        fun newInstance() = ActivityFragment()
+        fun newInstance() = ProfileFragment()
     }
 }
