+++ app/src/main/java/com/chesire/malime/flow/series/list/anime/AnimeAdapter.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.flow.overview.anime
+package com.chesire.malime.flow.series.list.anime
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
