+++ app/src/main/java/com/chesire/malime/flow/series/list/anime/AnimeFragment.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.flow.overview.anime
+package com.chesire.malime.flow.series.list.anime
 
 import android.os.Bundle
 import android.view.LayoutInflater
