+++ app/src/main/java/com/chesire/malime/flow/series/list/manga/MangaFragment.kt
@@ -1,27 +1,27 @@
-package com.chesire.malime.flow.overview.profile
+package com.chesire.malime.flow.series.list.manga
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.chesire.malime.databinding.FragmentProfileBinding
+import com.chesire.malime.databinding.FragmentMangaBinding
 import dagger.android.support.DaggerFragment
 
-class ProfileFragment : DaggerFragment() {
+class MangaFragment : DaggerFragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ) = FragmentProfileBinding.inflate(inflater, container, false).root
+    ) = FragmentMangaBinding.inflate(inflater, container, false).root
 
     companion object {
-        const val TAG = "ProfileFragment"
+        const val TAG = "MangaFragment"
 
         /**
          * Use this factory method to create a new instance of
          * this fragment using the provided parameters.
          *
-         * @return A new instance of fragment ProfileFragment.
+         * @return A new instance of fragment MangaFragment.
          */
-        fun newInstance() = ProfileFragment()
+        fun newInstance() = MangaFragment()
     }
 }
